Use setState updater functions and guard against null or undefined references.

+++ src/core_plugins/kibana/public/dashboard/panel/panel_options_menu.js
@@ -25,7 +25,10 @@ export class PanelOptionsMenu extends React.Component {
 
   onDeletePanel = () => {
     this.closePopover();
-    this.props.onDeletePanel();
+
+    if (this.props.onDeletePanel) {
+      this.props.onDeletePanel();
+    }
   };
 
   onToggleExpandPanel = () => {
